requirements.txt
cffi==1.9.1
Django==1.10.5
django-crispy-forms==1.6.0
django-cuser==2016.5.29
django-haystack==2.5.0
django-mptt==0.8.6
git+https://github.com/eedf/django-tagulous#egg=django-tagulous
django-tracking-fields==1.0.6
django-tracking2==0.3.2
elasticsearch==2.4.0
factory-boy==2.7.0
fake-factory==0.7.2
flake8==3.0.4
gunicorn==19.6.0
lxml==3.6.4
mccabe==0.5.2
mock==2.0.0
pbr==1.10.0
pep8==1.7.0
psycopg2==2.6.2
pycodestyle==2.0.0
pycparser==2.17
pyflakes==1.3.0
pylokit==0.8.1
python-dateutil==2.6.0
requests==2.11.1
requests-mock==1.1.0
six==1.10.0
git+https://github.com/eedf/templated-docs#egg=templated-docs
tika==1.13.1
urllib3==1.18
